Refactor LTI event listeners to be reusable
This PR adds monitoring of window messages to the wiki_page_show.js so
that LTI tools embedded in wiki pages can post messages to Canvas. In
particular this allows an embedded tool to resize it’s iframe which
solves problems when the content of the iframe changes its height. The
message handling is refactored into its own file. In addition, the
findDomForWindow method will l find the iframe that contains the window
that produced the window message so that the correct iframe is resized.
